refactor(ts): unify sidecar CLIs under src/cli/ with a composition root
Introduce a single CLI architecture under packages/ts/sidecar/src/cli/:

- CliCommand interface: run(argv) returns an exit code, never process.exit.
- CompositionRoot: the single production wiring point, composing
  PipelineFactory's pass graph with the Node IO adapters, FormatPipeline,
  reporters, and command classes (formatAllCommand, segmentPassCommand,
  fluentPassCommand, validateSyntaxCommand).
- One PassReporter and one SyntaxReporter, replacing FormatAllReporter,
  SyntaxErrorReporter, and the ad-hoc reporting loops in the blank-lines and
  fluent-chains entrypoints. Console output bytes are unchanged.
- FormatAllCommand owns the segment -> oxfmt -> fluent -> segment -> validate
  schedule; FormatPassCommand backs both standalone format passes;
  ValidateSyntaxCommand backs standalone validation.
- DTOs moved to cli/: PassCliDto, CliOptionsDto (cli/format-all-cli-dto),
  SyntaxCliDto. Flag grammar unchanged.
- Entry files (blank-lines, fluent-chains, validate-syntax) hold only main()
  plus the run-as-main guard.

Update sidecar.ts pipeline dispatch, the validate-syntax package script, and
the moved CLI tests to the new paths. Naming inversion is resolved: reporters
are named for what they report, not their former host module.
