+++ packages/ts/sidecar/src/cli/pass-cli-dto.ts
