+++ packages/ts/sidecar/src/sidecar.ts
@@ -76,7 +76,7 @@ switch (mode) {
 		// import; blank argv[1] so only the explicit main() call below runs.
 		process.argv[1] = '';
 
-		const { main } = await import('#sidecar/format-all');
+		const { main } = await import('#sidecar/cli/format-all');
 
 		await main();
